+++ forms-flow-bpm/src/main/java/org/camunda/bpm/extension/hooks/controllers/AdminController.java
@@ -6,29 +6,34 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.nimbusds.oauth2.sdk.util.CollectionUtils;
 import net.minidev.json.JSONArray;
 import org.apache.commons.lang.StringUtils;
-import org.camunda.bpm.extension.commons.connector.HTTPServiceInvoker;
+import org.camunda.bpm.engine.ProcessEngines;
 
+import org.camunda.bpm.extension.commons.connector.HTTPServiceInvoker;
 import org.camunda.bpm.extension.hooks.controllers.data.Authorization;
+import org.camunda.bpm.extension.hooks.controllers.data.AuthorizationInfo;
 import org.camunda.bpm.extension.hooks.controllers.data.AuthorizedAction;
-import org.camunda.bpm.extension.hooks.controllers.mapper.AuthorizationMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.oauth2.core.oidc.user.OidcUser;
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
 import org.camunda.bpm.engine.authorization.ProcessDefinitionPermissions;
 import org.camunda.bpm.engine.authorization.Resources;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 import javax.servlet.ServletException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -42,8 +47,8 @@ public class AdminController {
 
     private static final Logger LOGGER = Logger.getLogger(AdminController.class.getName());
 
-    @Autowired
-    private NamedParameterJdbcTemplate bpmJdbcTemplate;
+    @Value("${plugin.identity.keycloak.administratorGroupName}")
+    private String adminGroupName;
 
     @Autowired
     private HTTPServiceInvoker httpServiceInvoker;
@@ -51,9 +56,21 @@ public class AdminController {
     @Value("${formsflow.ai.api.url}")
     private String formsflowApiUrl;
 
-    @Value("${plugin.identity.keycloak.administratorGroupName}")
-    private String adminGroupName;
+    @GetMapping(value = "/engine-rest-ext/form/authorization", produces = MediaType.APPLICATION_JSON_VALUE)
+    private @ResponseBody AuthorizationInfo getFormAuthorization() throws ServletException {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        List<String> groups = getGroups(authentication);
+        AuthorizationInfo authorizationInfo = null;
 
+        if (CollectionUtils.isNotEmpty(groups) && groups.contains(adminGroupName)) {
+            authorizationInfo = new AuthorizationInfo(true, null);
+        } else {
+            authorizationInfo = new AuthorizationInfo(false, getAuthorization(groups));
+        }
+        return authorizationInfo;
+    }
+
+    @Deprecated
     @RequestMapping(value = "/engine-rest-ext/form", method = RequestMethod.GET, produces = "application/json")
     private @ResponseBody List<AuthorizedAction> getForms() throws ServletException {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
@@ -75,7 +92,6 @@ public class AdminController {
                             action.setFormName(formNode.get("formName").asText());
                             action.setProcessKey(formNode.get("processKey").asText());
                             formList.add(action);
-
                         }
                     }
 
@@ -159,18 +175,26 @@ public class AdminController {
 		return groupIds;
 	}
 
+
     /**
      * This method returns all authorization details of Groups.
      * @param groups
      * @return
      */
     private List<Authorization> getAuthorization(List<String> groups) {
-        String query = "select group_id_ groupid, user_id_ userid, resource_id_  resourceid from act_ru_authorization " +
-                "where resource_type_="+ Resources.PROCESS_DEFINITION.resourceType()+
-                " and perms_ >= " + ProcessDefinitionPermissions.CREATE_INSTANCE.getValue()  +
-                " and  group_id_ IN (:groups) ";
-        MapSqlParameterSource parameters = new MapSqlParameterSource();
-        parameters.addValue("groups", groups);
-        return bpmJdbcTemplate.query(query, parameters, new AuthorizationMapper());
+
+        List<Authorization> authorizationList = new ArrayList<>();
+
+        String[] groupIds = groups.size() > 0 ? groups.toArray(new String[0]) : new String[]{};
+        List<org.camunda.bpm.engine.authorization.Authorization> authorizations =  ProcessEngines.getDefaultProcessEngine().getAuthorizationService().createAuthorizationQuery()
+                .resourceType(Resources.PROCESS_DEFINITION.resourceType())
+                .hasPermission(ProcessDefinitionPermissions.CREATE_INSTANCE)
+                .groupIdIn(groupIds).list();
+
+        authorizations.forEach(authorization -> {
+            Authorization auth = new Authorization(authorization.getGroupId(), authorization.getUserId(), authorization.getResourceId());
+            authorizationList.add(auth);
+        });
+        return authorizationList;
     }
 }